+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/nfts/page.tsx
@@ -1,8 +1,8 @@
 import { notFound, redirect } from "next/navigation";
-import { ContractNFTPage } from "../../../../../../contract-ui/tabs/nfts/page";
 import { resolveFunctionSelectors } from "../../../../../../lib/selectors";
 import { getContractPageParamsInfo } from "../_utils/getContractFromParams";
 import { getContractPageMetadata } from "../_utils/getContractPageMetadata";
+import { ContractNFTPage } from "./ContractNFTPage";
 
 export default async function Page(props: {
   params: {
